Extract SSHConnectionValidator to ssh_connection_validator.py
Move the form-input validation classes out of the connection_dialog.py
god-object into a new leaf module, sshpilot/ssh_connection_validator.py:
ValidationResult and SSHConnectionValidator (connection name / hostname / IP /
port / username validation, plus a local ssh-keygen -y passphrase check).

These validate user-entered fields and never open an SSH connection, so they are
unrelated to the single connection/auth path. They have no GTK dependency, so a
leaf module is their natural home and makes them testable in isolation.

connection_dialog.py re-exports both names (it uses SSHConnectionValidator and
the ValidationResult type hint internally, and existing suites import
SSHConnectionValidator from connection_dialog) so back-compat is preserved.

Pure code move, no behaviour change. Adds tests/test_ssh_connection_validator.py
(19 cases) covering the validation rules directly — they had no focused coverage
before.

Verify: ruff clean; all 106 submodules import; full unit suite 1071 passed
(1052 + 19 new), 0 failed.
